+++ diary/urls/diary_urls.py
@@ -1,6 +1,8 @@
 from django.urls import path
 
-from diary.views.diary_views import *
+from diary.views.diary_query_views import *
+from diary.views.diray_command_views import *
+from diary.views.graph_views import GetNodeData
 from diary.views.quiz_views import CheckAnswerView
 
 urlpatterns = [
